Make KafkaApp responsible of deploying/cleaning streams bootstrap components (#522)
