+++ docs/docs/resources/pipeline-components/dependencies/defaults_pipeline_component_dependencies.yaml
@@ -5,7 +5,6 @@ kafka-app.yaml:
 - prefix.yaml
 - from_.yaml
 - to.yaml
-- values-kafka-app.yaml
 kafka-connector.yaml:
 - prefix.yaml
 - from_.yaml
