+++ docs/docs/resources/pipeline-components/kafka-app.yaml
@@ -44,9 +44,3 @@
           cleanup.policy: compact
     models: # SchemaProvider is initiated with the values given here
       model: model
-  values: # required
-    streams: # required
-      brokers: ${config.kafka_brokers} # required
-      schemaRegistryUrl: ${config.schema_registry.url}
-    nameOverride: override-with-this-name # kafka-app-specific
-    imageTag: "1.0.0" # Example values that are shared between streams-app and producer-app
