+++ docs/docs/resources/pipeline-components/pipeline.yaml
@@ -105,12 +105,6 @@
           cleanup.policy: compact
     models: # SchemaProvider is initiated with the values given here
       model: model
-  values: # required
-    streams: # required
-      brokers: ${config.kafka_brokers} # required
-      schemaRegistryUrl: ${config.schema_registry.url}
-    nameOverride: override-with-this-name # kafka-app-specific
-    imageTag: "1.0.0" # Example values that are shared between streams-app and producer-app
 # Kafka sink connector
 - type: kafka-sink-connector
   name: kafka-sink-connector # required
