+++ docs/docs/resources/pipeline-defaults/defaults-kafka-app.yaml
@@ -45,9 +45,3 @@ kafka-app:
           cleanup.policy: compact
     models: # SchemaProvider is initiated with the values given here
       model: model
-  values: # required
-    streams: # required
-      brokers: ${config.kafka_brokers} # required
-      schemaRegistryUrl: ${config.schema_registry.url}
-    nameOverride: override-with-this-name # kafka-app-specific
-    imageTag: "1.0.0" # Example values that are shared between streams-app and producer-app
