+++ kpops/components/common/streams_bootstrap.py
@@ -2,14 +2,21 @@ from __future__ import annotations
 
 import logging
 from abc import ABC
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Any
 
 import pydantic
-from pydantic import Field
+from pydantic import AliasChoices, ConfigDict, Field
 
 from kpops.component_handlers.helm_wrapper.model import HelmRepoConfig
 from kpops.components.base_components.helm_app import HelmApp, HelmAppValues
+from kpops.components.common.topic import KafkaTopic, KafkaTopicStr
 from kpops.utils.docstring import describe_attr
+from kpops.utils.pydantic import (
+    CamelCaseConfigModel,
+    DescConfigModel,
+    exclude_by_value,
+    exclude_defaults,
+)
 
 if TYPE_CHECKING:
     try:
@@ -29,6 +36,64 @@ IMAGE_TAG_PATTERN = r"^[a-zA-Z0-9_][a-zA-Z0-9._-]{0,127}$"
 log = logging.getLogger("StreamsBootstrap")
 
 
+class KafkaStreamsConfig(CamelCaseConfigModel, DescConfigModel):
+    """Kafka Streams config.
+
+    :param brokers: Brokers
+    :param schema_registry_url: URL of the schema registry, defaults to None
+    :param extra_output_topics: Extra output topics
+    :param output_topic: Output topic, defaults to None
+    """
+
+    brokers: str = Field(default=..., description=describe_attr("brokers", __doc__))
+    schema_registry_url: str | None = Field(
+        default=None,
+        validation_alias=AliasChoices(
+            "schema_registry_url", "schemaRegistryUrl"
+        ),  # TODO: same for other camelcase fields, avoids duplicates during enrichment
+        description=describe_attr("schema_registry_url", __doc__),
+    )
+    extra_output_topics: dict[str, KafkaTopicStr] = Field(
+        default={}, description=describe_attr("extra_output_topics", __doc__)
+    )
+    output_topic: KafkaTopicStr | None = Field(
+        default=None,
+        description=describe_attr("output_topic", __doc__),
+        json_schema_extra={},
+    )
+
+    model_config = ConfigDict(extra="allow")
+
+    @pydantic.field_validator("extra_output_topics", mode="before")
+    @classmethod
+    def deserialize_extra_output_topics(
+        cls, extra_output_topics: dict[str, str] | Any
+    ) -> dict[str, KafkaTopic] | Any:
+        if isinstance(extra_output_topics, dict):
+            return {
+                role: KafkaTopic(name=topic_name)
+                for role, topic_name in extra_output_topics.items()
+            }
+        return extra_output_topics
+
+    @pydantic.field_serializer("extra_output_topics")
+    def serialize_extra_output_topics(
+        self, extra_topics: dict[str, KafkaTopic]
+    ) -> dict[str, str]:
+        return {role: topic.name for role, topic in extra_topics.items()}
+
+    # TODO(Ivan Yordanov): Currently hacky and potentially unsafe. Find cleaner solution
+    @pydantic.model_serializer(mode="wrap", when_used="always")
+    def serialize_model(
+        self,
+        default_serialize_handler: pydantic.SerializerFunctionWrapHandler,
+        info: pydantic.SerializationInfo,
+    ) -> dict[str, Any]:
+        return exclude_defaults(
+            self, exclude_by_value(default_serialize_handler(self), None)
+        )
+
+
 class StreamsBootstrapValues(HelmAppValues):
     """Base value class for all streams bootstrap related components.
 
@@ -39,6 +104,10 @@ class StreamsBootstrapValues(HelmAppValues):
         default="latest",
         pattern=IMAGE_TAG_PATTERN,
         description=describe_attr("image_tag", __doc__),
+    )
+
+    streams: KafkaStreamsConfig = Field(
+        default=..., description=describe_attr("streams", __doc__)
     )
