+++ kpops/components/streams_bootstrap/producer/model.py
@@ -1,10 +1,9 @@
 from pydantic import ConfigDict, Field
 
-from kpops.components.base_components.kafka_app import (
-    KafkaAppValues,
+from kpops.components.common.streams_bootstrap import (
     KafkaStreamsConfig,
+    StreamsBootstrapValues,
 )
-from kpops.components.common.streams_bootstrap import StreamsBootstrapValues
 from kpops.utils.docstring import describe_attr
 
 
@@ -12,7 +11,7 @@ class ProducerStreamsConfig(KafkaStreamsConfig):
     """Kafka Streams settings specific to Producer."""
 
 
-class ProducerAppValues(StreamsBootstrapValues, KafkaAppValues):
+class ProducerAppValues(StreamsBootstrapValues):
     """Settings specific to producers.
 
     :param streams: Kafka Streams settings
